test LAPACKE ILP64

+++ clients/common/cblas_interface.cpp
@@ -3566,6 +3566,7 @@ void ref_syr2k(hipblasFillMode_t     uplo,
 // syrkx
 // Use syrk with A == B for now.
 
+/*
 // trsm
 template <>
 void ref_trsm<float>(hipblasSideMode_t  side,
@@ -3676,6 +3677,8 @@ void ref_trsm<hipblasDoubleComplex>(hipblasSideMode_t           side,
                 B,
                 ldb);
 }
+
+*/
 
 // trtri
 template <>
@@ -3822,36 +3825,36 @@ void ref_trmm<hipblasDoubleComplex>(hipblasSideMode_t           side,
 
 // potrf
 template <>
-int ref_potrf(char uplo, int m, float* A, int lda)
+int64_t ref_potrf(char uplo, int64_t m, float* A, int64_t lda)
 {
-    int info;
+    int64_t info;
     //spotrf_(&uplo, &m, A, &lda, &info);
     info = LAPACKE_spotrf(LAPACK_COL_MAJOR, uplo, m, A, lda);
     return info;
 }
 
 template <>
-int ref_potrf(char uplo, int m, double* A, int lda)
+int64_t ref_potrf(char uplo, int64_t m, double* A, int64_t lda)
 {
-    int info;
+    int64_t info;
     //dpotrf_(&uplo, &m, A, &lda, &info);
     info = LAPACKE_dpotrf(LAPACK_COL_MAJOR, uplo, m, A, lda);
     return info;
 }
 
 template <>
-int ref_potrf(char uplo, int m, hipblasComplex* A, int lda)
+int64_t ref_potrf(char uplo, int64_t m, hipblasComplex* A, int64_t lda)
 {
-    int info;
+    int64_t info;
     //cpotrf_(&uplo, &m, A, &lda, &info);
     info = LAPACKE_cpotrf(LAPACK_COL_MAJOR, uplo, m, (lapack_complex_float*)A, lda);
     return info;
 }
 
 template <>
-int ref_potrf(char uplo, int m, hipblasDoubleComplex* A, int lda)
+int64_t ref_potrf(char uplo, int64_t m, hipblasDoubleComplex* A, int64_t lda)
 {
-    int info;
+    int64_t info;
     //zpotrf_(&uplo, &m, A, &lda, &info);
     info = LAPACKE_zpotrf(LAPACK_COL_MAJOR, uplo, m, (lapack_complex_double*)A, lda);
     return info;
@@ -3859,36 +3862,38 @@ int ref_potrf(char uplo, int m, hipblasDoubleComplex* A, int lda)
 
 // getrf
 template <>
-int ref_getrf<float>(int m, int n, float* A, int lda, int* ipiv)
+int64_t ref_getrf<float>(int64_t m, int64_t n, float* A, int64_t lda, int64_t* ipiv)
 {
-    int info;
+    int64_t info;
     //sgetrf_(&m, &n, A, &lda, ipiv, &info);
     info = LAPACKE_sgetrf(LAPACK_COL_MAJOR, m, n, A, lda, ipiv);
     return info;
 }
 
 template <>
-int ref_getrf<double>(int m, int n, double* A, int lda, int* ipiv)
+int64_t ref_getrf<double>(int64_t m, int64_t n, double* A, int64_t lda, int64_t* ipiv)
 {
-    int info;
+    int64_t info;
     //dgetrf_(&m, &n, A, &lda, ipiv, &info);
     LAPACKE_dgetrf(LAPACK_COL_MAJOR, m, n, A, lda, ipiv);
     return info;
 }
 
 template <>
-int ref_getrf<hipblasComplex>(int m, int n, hipblasComplex* A, int lda, int* ipiv)
+int64_t
+    ref_getrf<hipblasComplex>(int64_t m, int64_t n, hipblasComplex* A, int64_t lda, int64_t* ipiv)
 {
-    int info;
+    int64_t info;
     //cgetrf_(&m, &n,  (lapack_complex_float*)A, &lda, ipiv, &info);
     LAPACKE_cgetrf(LAPACK_COL_MAJOR, m, n, (lapack_complex_float*)A, lda, ipiv);
     return info;
 }
 
 template <>
-int ref_getrf<hipblasDoubleComplex>(int m, int n, hipblasDoubleComplex* A, int lda, int* ipiv)
+int64_t ref_getrf<hipblasDoubleComplex>(
+    int64_t m, int64_t n, hipblasDoubleComplex* A, int64_t lda, int64_t* ipiv)
 {
-    int info;
+    int64_t info;
     //zgetrf_(&m, &n,  (lapack_complex_double*)A, &lda, ipiv, &info);
     LAPACKE_zgetrf(LAPACK_COL_MAJOR, m, n, (lapack_complex_double*)A, lda, ipiv);
     return info;
@@ -3896,29 +3901,48 @@ int ref_getrf<hipblasDoubleComplex>(int m, int n, hipblasDoubleComplex* A, int l
 
 // getrs
 template <>
-int ref_getrs<float>(char trans, int n, int nrhs, float* A, int lda, int* ipiv, float* B, int ldb)
+int64_t ref_getrs<float>(char     trans,
+                         int64_t  n,
+                         int64_t  nrhs,
+                         float*   A,
+                         int64_t  lda,
+                         int64_t* ipiv,
+                         float*   B,
+                         int64_t  ldb)
 {
-    int info;
+    int64_t info;
     //sgetrs_(&trans, &n, &nrhs, A, &lda, ipiv, B, &ldb, &info, 1);
     info = LAPACKE_sgetrs(LAPACK_COL_MAJOR, trans, n, nrhs, A, lda, ipiv, B, ldb);
     return info;
 }
 
 template <>
-int ref_getrs<double>(
-    char trans, int n, int nrhs, double* A, int lda, int* ipiv, double* B, int ldb)
+int64_t ref_getrs<double>(char     trans,
+                          int64_t  n,
+                          int64_t  nrhs,
+                          double*  A,
+                          int64_t  lda,
+                          int64_t* ipiv,
+                          double*  B,
+                          int64_t  ldb)
 {
-    int info;
+    int64_t info;
     //dgetrs_(&trans, &n, &nrhs, A, &lda, ipiv, B, &ldb, &info, 1);
     info = LAPACKE_dgetrs(LAPACK_COL_MAJOR, trans, n, nrhs, A, lda, ipiv, B, ldb);
     return info;
 }
 
 template <>
-int ref_getrs<hipblasComplex>(
-    char trans, int n, int nrhs, hipblasComplex* A, int lda, int* ipiv, hipblasComplex* B, int ldb)
+int64_t ref_getrs<hipblasComplex>(char            trans,
+                                  int64_t         n,
+                                  int64_t         nrhs,
+                                  hipblasComplex* A,
+                                  int64_t         lda,
+                                  int64_t*        ipiv,
+                                  hipblasComplex* B,
+                                  int64_t         ldb)
 {
-    int info;
+    int64_t info;
     // cgetrs_(&trans,
     //         &n,
     //         &nrhs,
@@ -3942,16 +3966,16 @@ int ref_getrs<hipblasComplex>(
 }
 
 template <>
-int ref_getrs<hipblasDoubleComplex>(char                  trans,
-                                    int                   n,
-                                    int                   nrhs,
-                                    hipblasDoubleComplex* A,
-                                    int                   lda,
-                                    int*                  ipiv,
-                                    hipblasDoubleComplex* B,
-                                    int                   ldb)
+int64_t ref_getrs<hipblasDoubleComplex>(char                  trans,
+                                        int64_t               n,
+                                        int64_t               nrhs,
+                                        hipblasDoubleComplex* A,
+                                        int64_t               lda,
+                                        int64_t*              ipiv,
+                                        hipblasDoubleComplex* B,
+                                        int64_t               ldb)
 {
-    int info;
+    int64_t info;
     // zgetrs_(&trans,
     //         &n,
     //         &nrhs,
@@ -3976,38 +4000,44 @@ int ref_getrs<hipblasDoubleComplex>(char                  trans,
 
 // getri
 template <>
-int ref_getri<float>(int n, float* A, int lda, int* ipiv, float* work, int lwork)
+int64_t
+    ref_getri<float>(int64_t n, float* A, int64_t lda, int64_t* ipiv, float* work, int64_t lwork)
 {
-    int info;
+    int64_t info;
     //sgetri_(&n, A, &lda, ipiv, work, &lwork, &info);
     info = LAPACKE_sgetri(LAPACK_COL_MAJOR, n, A, lda, ipiv);
     return info;
 }
 
 template <>
-int ref_getri<double>(int n, double* A, int lda, int* ipiv, double* work, int lwork)
+int64_t
+    ref_getri<double>(int64_t n, double* A, int64_t lda, int64_t* ipiv, double* work, int64_t lwork)
 {
-    int info;
+    int64_t info;
     //dgetri_(&n, A, &lda, ipiv, work, &lwork, &info);
     info = LAPACKE_dgetri(LAPACK_COL_MAJOR, n, A, lda, ipiv);
     return info;
 }
 
 template <>
-int ref_getri<hipblasComplex>(
-    int n, hipblasComplex* A, int lda, int* ipiv, hipblasComplex* work, int lwork)
+int64_t ref_getri<hipblasComplex>(
+    int64_t n, hipblasComplex* A, int64_t lda, int64_t* ipiv, hipblasComplex* work, int64_t lwork)
 {
-    int info;
+    int64_t info;
     //cgetri_(&n, A, &lda, ipiv, work, &lwork, &info);
     info = LAPACKE_cgetri(LAPACK_COL_MAJOR, n, (lapack_complex_float*)A, lda, ipiv);
     return info;
 }
 
 template <>
-int ref_getri<hipblasDoubleComplex>(
-    int n, hipblasDoubleComplex* A, int lda, int* ipiv, hipblasDoubleComplex* work, int lwork)
+int64_t ref_getri<hipblasDoubleComplex>(int64_t               n,
+                                        hipblasDoubleComplex* A,
+                                        int64_t               lda,
+                                        int64_t*              ipiv,
+                                        hipblasDoubleComplex* work,
+                                        int64_t               lwork)
 {
-    int info;
+    int64_t info;
     //zgetri_(&n, A, &lda, ipiv, work, &lwork, &info);
     info = LAPACKE_zgetri(LAPACK_COL_MAJOR, n, (lapack_complex_double*)A, lda, ipiv);
     return info;
@@ -4015,27 +4045,34 @@ int ref_getri<hipblasDoubleComplex>(
 
 // geqrf
 template <>
-int ref_geqrf<float>(int m, int n, float* A, int lda, float* tau, float* work, int lwork)
+int64_t ref_geqrf<float>(
+    int64_t m, int64_t n, float* A, int64_t lda, float* tau, float* work, int64_t lwork)
 {
-    int info;
+    int64_t info;
     //sgeqrf_(&m, &n, A, &lda, tau, work, &lwork, &info);
     info = LAPACKE_sgeqrf(LAPACK_COL_MAJOR, m, n, A, lda, tau);
     return info;
 }
 
 template <>
-int ref_geqrf<double>(int m, int n, double* A, int lda, double* tau, double* work, int lwork)
+int64_t ref_geqrf<double>(
+    int64_t m, int64_t n, double* A, int64_t lda, double* tau, double* work, int64_t lwork)
 {
-    int info;
+    int64_t info;
     //dgeqrf_(&m, &n, A, &lda, tau, work, &lwork, &info);
     info = LAPACKE_dgeqrf(LAPACK_COL_MAJOR, m, n, A, lda, tau);
     return info;
 }
 template <>
-int ref_geqrf<hipblasComplex>(
-    int m, int n, hipblasComplex* A, int lda, hipblasComplex* tau, hipblasComplex* work, int lwork)
+int64_t ref_geqrf<hipblasComplex>(int64_t         m,
+                                  int64_t         n,
+                                  hipblasComplex* A,
+                                  int64_t         lda,
+                                  hipblasComplex* tau,
+                                  hipblasComplex* work,
+                                  int64_t         lwork)
 {
-    int info;
+    int64_t info;
     //cgeqrf_(&m, &n, A, &lda, tau, work, &lwork, &info);
     info = LAPACKE_cgeqrf(
         LAPACK_COL_MAJOR, m, n, (lapack_complex_float*)A, lda, (lapack_complex_float*)tau);
@@ -4043,15 +4080,15 @@ int ref_geqrf<hipblasComplex>(
 }
 
 template <>
-int ref_geqrf<hipblasDoubleComplex>(int                   m,
-                                    int                   n,
-                                    hipblasDoubleComplex* A,
-                                    int                   lda,
-                                    hipblasDoubleComplex* tau,
-                                    hipblasDoubleComplex* work,
-                                    int                   lwork)
+int64_t ref_geqrf<hipblasDoubleComplex>(int64_t               m,
+                                        int64_t               n,
+                                        hipblasDoubleComplex* A,
+                                        int64_t               lda,
+                                        hipblasDoubleComplex* tau,
+                                        hipblasDoubleComplex* work,
+                                        int64_t               lwork)
 {
-    int info;
+    int64_t info;
     //zgeqrf_(&m, &n, A, &lda, tau, work, &lwork, &info);
     info = LAPACKE_zgeqrf(
         LAPACK_COL_MAJOR, m, n, (lapack_complex_double*)A, lda, (lapack_complex_double*)tau);
@@ -4060,54 +4097,54 @@ int ref_geqrf<hipblasDoubleComplex>(int                   m,
 
 // gels
 template <>
-int ref_gels<float>(char   trans,
-                    int    m,
-                    int    n,
-                    int    nrhs,
-                    float* A,
-                    int    lda,
-                    float* B,
-                    int    ldb,
-                    float* work,
-                    int    lwork)
+int64_t ref_gels<float>(char    trans,
+                        int64_t m,
+                        int64_t n,
+                        int64_t nrhs,
+                        float*  A,
+                        int64_t lda,
+                        float*  B,
+                        int64_t ldb,
+                        float*  work,
+                        int64_t lwork)
 {
-    int info;
+    int64_t info;
     //sgels_(&trans, &m, &n, &nrhs, A, &lda, B, &ldb, work, &lwork, &info);
     info = LAPACKE_sgels(LAPACK_COL_MAJOR, trans, m, n, nrhs, A, lda, B, ldb);
     return info;
 }
 
 template <>
-int ref_gels<double>(char    trans,
-                     int     m,
-                     int     n,
-                     int     nrhs,
-                     double* A,
-                     int     lda,
-                     double* B,
-                     int     ldb,
-                     double* work,
-                     int     lwork)
+int64_t ref_gels<double>(char    trans,
+                         int64_t m,
+                         int64_t n,
+                         int64_t nrhs,
+                         double* A,
+                         int64_t lda,
+                         double* B,
+                         int64_t ldb,
+                         double* work,
+                         int64_t lwork)
 {
-    int info;
+    int64_t info;
     //dgels_(&trans, &m, &n, &nrhs, A, &lda, B, &ldb, work, &lwork, &info);
     info = LAPACKE_dgels(LAPACK_COL_MAJOR, trans, m, n, nrhs, A, lda, B, ldb);
     return info;
 }
 
 template <>
-int ref_gels<hipblasComplex>(char            trans,
-                             int             m,
-                             int             n,
-                             int             nrhs,
-                             hipblasComplex* A,
-                             int             lda,
-                             hipblasComplex* B,
-                             int             ldb,
-                             hipblasComplex* work,
-                             int             lwork)
+int64_t ref_gels<hipblasComplex>(char            trans,
+                                 int64_t         m,
+                                 int64_t         n,
+                                 int64_t         nrhs,
+                                 hipblasComplex* A,
+                                 int64_t         lda,
+                                 hipblasComplex* B,
+                                 int64_t         ldb,
+                                 hipblasComplex* work,
+                                 int64_t         lwork)
 {
-    int info;
+    int64_t info;
     //cgels_(&trans, &m, &n, &nrhs, A, &lda, B, &ldb, work, &lwork, &info);
     info = LAPACKE_cgels(LAPACK_COL_MAJOR,
                          trans,
@@ -4122,18 +4159,18 @@ int ref_gels<hipblasComplex>(char            trans,
 }
 
 template <>
-int ref_gels<hipblasDoubleComplex>(char                  trans,
-                                   int                   m,
-                                   int                   n,
-                                   int                   nrhs,
-                                   hipblasDoubleComplex* A,
-                                   int                   lda,
-                                   hipblasDoubleComplex* B,
-                                   int                   ldb,
-                                   hipblasDoubleComplex* work,
-                                   int                   lwork)
+int64_t ref_gels<hipblasDoubleComplex>(char                  trans,
+                                       int64_t               m,
+                                       int64_t               n,
+                                       int64_t               nrhs,
+                                       hipblasDoubleComplex* A,
+                                       int64_t               lda,
+                                       hipblasDoubleComplex* B,
+                                       int64_t               ldb,
+                                       hipblasDoubleComplex* work,
+                                       int64_t               lwork)
 {
-    int info;
+    int64_t info;
     //zgels_(&trans, &m, &n, &nrhs, A, &lda, B, &ldb, work, &lwork, &info);
     info = LAPACKE_zgels(LAPACK_COL_MAJOR,
                          trans,
